chore: rename http getter package (#1349)
