+++ plugin/http-getter/image_test.go
